Removed OCR class because it is not needed
