+++ src/TikaOnDotnet.TextExtractor/TikaOnDotnet.TextExtraction.csproj
@@ -54,7 +54,6 @@
     <Compile Include="TextExtractionException.cs" />
     <Compile Include="TextExtractionResult.cs" />
     <Compile Include="TextExtractor.cs" />
-    <Compile Include="TextExtractorOcr.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
